Cleaned up MotionFunctions
