--- Assets/Scripts/Player/MotionFunctions.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: de9d200144e72164698346b44c2ad6b2
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
